--- src/context/categoryContext.jsx
@@ -1,26 +1,0 @@
-import { createContext, useState, useEffect } from "react"
-
-import { getCategoriesAndDocuments } from "../utils/firebase/FirebaseUtils"
-
-export const CategoriesContext = createContext({
-  categoriesMap: {},
-})
-
-export const CategoriesProvider = ({ children }) => {
-  const [categoriesMap, setCategoriesMap] = useState({})
-
-  useEffect(() => {
-    const getCategoriesMap = async () => {
-      const categoryMap = await getCategoriesAndDocuments()
-      setCategoriesMap(categoryMap)
-    }
-    getCategoriesMap()
-  }, [])
-
-  const value = { categoriesMap }
-  return (
-    <CategoriesContext.Provider value={value}>
-      {children}
-    </CategoriesContext.Provider>
-  )
-}
